This patch is an indirect workaround to what seems to be a bug in kdepim's vcard support.
WARNING: This causes it to diverge from the official spec, so this might not be necessary for (or it may even hurt) syncing with other endpoints such as Evolution.

Some details:
Section 2.5 of RFC2426 claims that commas must be escaped in Structured Type Values. This comes into play with recurrence rules. For example, if there is an event that recurs weekly on Saturday and Sunday, the corresponding rule would be:
    "FREQ=WEEKLY;BYDAY=SU\,SA"
Unfortunately, kdepim (I'm not sure if it is libical's or libkcal's fault) reads and writes rules without the escaped comma:
    "FREQ=WEEKLY;BYDAY=SU,SA"
To this end, this patch causes OpenSync to write out recurrence rules without the '\' so that kdepim will understand them.