Patch from Ian Abbott, thanks!
pls sometimes frees memory twice
If you run pls with a relative path and a trailing slash, e.g.:

$ pls Business/

The memory pointed to by the 'wide_path' variable gets freed twice - once
by the call to 'adjust_remote_path' in 'list_matching_files' and once in
'main'.

The attached patch fixes it by removing the call to 'adjust_remote_path' in
'list_matching_files' and adding a call to 'adjust_remote_path' in the
"trailing backslash" case in 'main'.